data/requirements.txt
txt

qdrant-client
langchain
langchain-community
langchain-openai
langchain-qdrant
openai
PyPDF2
python-dotenv